+++ src/v2/providers/alerts/alerts.ts
@@ -23,7 +23,7 @@
 import { ManifestEndpoint } from "../../../runtime/manifest";
 import { CloudEvent, CloudFunction } from "../../core";
 import * as options from "../../options";
-import { Expression } from "../../params";
+import { Expression } from "../../../params";
 
 /**
  * The CloudEvent data emitted by Firebase Alerts.
