+++ src/v2/providers/alerts/appDistribution.ts
@@ -28,7 +28,7 @@
 import { CloudEvent, CloudFunction } from "../../core";
 import * as options from "../../options";
 import { FirebaseAlertData, getEndpointAnnotation } from "./alerts";
-import { Expression } from "../../params";
+import { Expression } from "../../../params";
 
 /**
  * The internal payload object for adding a new tester device to app distribution.
